Refactor color package
