+++ color/logger.go
@@ -1,9 +1,4 @@
-// Package color contains goyek features which additionally
-// have colors.
-//
-// Set NO_COLOR environment variable to a non-empty string
-// or use the NoColor function to prevent colorizing the output.
-package color
+
 
 import (
 	"fmt"
@@ -17,59 +12,6 @@ import (
 	"github.com/fatih/color"
 	"github.com/goyek/goyek/v2"
 )
-
-func init() {
-	color.NoColor = os.Getenv("NO_COLOR") != ""
-}
-
-// NoColor prevents colorizing the output.
-func NoColor() {
-	color.NoColor = true
-}
-
-// ReportStatus is a middleware which reports the task run status with colors.
-//
-// The format is based on the reports provided by the Go test runner.
-func ReportStatus(next goyek.Runner) goyek.Runner {
-	return func(in goyek.Input) goyek.Result {
-		c := color.New(color.FgBlue)
-
-		// report start task
-		c.Fprintf(in.Output, "===== TASK  %s\n", in.TaskName)
-		start := time.Now()
-
-		// run
-		res := next(in)
-
-		// report task end
-		c = color.New(color.FgGreen)
-		status := "PASS"
-		switch res.Status {
-		case goyek.StatusFailed:
-			c = color.New(color.FgRed, color.Bold)
-			status = "FAIL"
-		case goyek.StatusSkipped:
-			c = color.New(color.FgYellow)
-			status = "SKIP"
-		case goyek.StatusNotRun:
-			status = "NOOP"
-		}
-		c.Fprintf(in.Output, "----- %s: %s (%.2fs)\n", status, in.TaskName, time.Since(start).Seconds())
-
-		// report panic if happened
-		if res.PanicStack != nil {
-			if res.PanicValue != nil {
-				c.Fprintf(in.Output, "panic: %v", res.PanicValue)
-			} else {
-				c.Fprint(in.Output, "panic(nil) or runtime.Goexit() called")
-			}
-			io.WriteString(in.Output, "\n\n") //nolint:errcheck // not checking errors when writing to output
-			c.Fprintf(in.Output, "%s", res.PanicStack)
-		}
-
-		return res
-	}
-}
 
 // CodeLineLogger decorates the log with code line information, identation and colors.
 type CodeLineLogger struct {
